requirements.txt
chromadb
cohere
fastapi
firebase_admin
fireworks-ai
gunicorn
json_repair
openai
PyJWT
requests